Updates dependencies and fixes route encoding
Updates various dependencies in the web-app to their latest versions, including react-router, lucide-react, storybook packages, eslint packages, sass and vite.

Removes unnecessary URI encoding in the getUserProfileRoute function, simplifying the route generation process.

Adjusts the return type of `getAuthToken`, `getLocale`, and `getTheme` in repositories to remove the explicit undefined type.
This simplifies type handling where these values are used.

Moves the import statement for `Fallback` in the `ErrorBoundary` component to the top of the file for better code readability.